Update base-level description to inlucde input and output fields descriptions
